--- src/umbp/distributed/common/types.cpp
@@ -1,55 +1,0 @@
-// Copyright © Advanced Micro Devices, Inc. All rights reserved.
-//
-// MIT License
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-#include "umbp/common/types.h"
-
-namespace mori::umbp {
-
-bool Location::operator==(const Location& other) const {
-  return node_id == other.node_id && location_id == other.location_id && size == other.size &&
-         tier == other.tier;
-}
-
-const char* TierTypeName(TierType t) {
-  switch (t) {
-    case TierType::HBM:
-      return "HBM";
-    case TierType::DRAM:
-      return "DRAM";
-    case TierType::SSD:
-      return "SSD";
-    default:
-      return "UNKNOWN";
-  }
-}
-
-const char* ClientStatusName(ClientStatus s) {
-  switch (s) {
-    case ClientStatus::ALIVE:
-      return "ALIVE";
-    case ClientStatus::EXPIRED:
-      return "EXPIRED";
-    default:
-      return "UNKNOWN";
-  }
-}
-
-}  // namespace mori::umbp
