+++ src/umbp/include/umbp/common/storage_tier.h
@@ -23,6 +23,8 @@
 
 #include <cstddef>
 
+namespace mori::umbp {
+
 enum class StorageTier : int { CPU_DRAM = 0, LOCAL_SSD = 1 };
 
 struct LocalLocation {
@@ -30,3 +32,5 @@ struct LocalLocation {
   size_t offset;
   size_t size;
 };
+
+}  // namespace mori::umbp
